+++ routes/index.js
@@ -1,0 +1,4 @@
+module.exports.getUser = require('./get')
+module.exports.addUser = require('./post')
+module.exports.deleteUser = require('./delete')
+module.exports.editUser = require('./put')
